chore: 글로벌 entity 이동
