+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/feature/testcase/dto/response/TestcaseGenerationResponse.java
@@ -17,7 +17,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class TestcaseGenerationResponse {
     private String processedAt;             // 처리 완료 시간
-    private double validationRate;          // 검증률
+    private Double validationRate;          // 검증률
 
     private List<TestcaseDataDto> tcList;     // 테스트 데이터 목록 (parameter 구조 + 예상값)
 }
