+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/feature/testcase/repository/ApiListRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.sk.skala.axcalibur.feature.testcase.entity.ApiListEntity;
+import com.sk.skala.axcalibur.global.entity.ApiListEntity;
 
 public interface ApiListRepository extends JpaRepository<ApiListEntity, Integer> {
     List<ApiListEntity> findByIdIn(List<Integer> ids);
