+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/feature/testcase/repository/CategoryRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.sk.skala.axcalibur.feature.testcase.entity.CategoryEntity;
+import com.sk.skala.axcalibur.global.entity.CategoryEntity;
 
 public interface CategoryRepository extends JpaRepository<CategoryEntity, Integer> {
     Optional<CategoryEntity> findByName(String name);
