+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/feature/testcase/repository/ContextRepository.java
@@ -4,7 +4,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.sk.skala.axcalibur.feature.testcase.entity.ContextEntity;
+import com.sk.skala.axcalibur.global.entity.ContextEntity;
 
 public interface ContextRepository extends JpaRepository<ContextEntity, Integer> {
     Optional<ContextEntity> findByName(String name);
