+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/feature/testcase/repository/MappingRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.sk.skala.axcalibur.feature.testcase.entity.MappingEntity;
+import com.sk.skala.axcalibur.global.entity.MappingEntity;
 
 public interface MappingRepository extends JpaRepository<MappingEntity, Integer> {
     List<MappingEntity> findByScenarioKey_Id(Integer scenarioId);
