+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/feature/testcase/repository/ParameterRepository.java
@@ -5,7 +5,7 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.sk.skala.axcalibur.feature.testcase.entity.ParameterEntity;
+import com.sk.skala.axcalibur.global.entity.ParameterEntity;
 
 public interface ParameterRepository extends JpaRepository<ParameterEntity, Integer> {
     List<ParameterEntity> findByApiListKey_Id(Integer apiId);
