+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/feature/testcase/repository/ProjectRepository.java
@@ -2,6 +2,6 @@ package com.sk.skala.axcalibur.feature.testcase.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.sk.skala.axcalibur.feature.testcase.entity.ProjectEntity;
+import com.sk.skala.axcalibur.global.entity.ProjectEntity;
 
 public interface ProjectRepository extends JpaRepository<ProjectEntity, Integer> {}
