+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/feature/testcase/service/TcGeneratorServiceImpl.java
@@ -18,9 +18,6 @@ import com.sk.skala.axcalibur.feature.testcase.dto.request.TcRequestPayload;
 import com.sk.skala.axcalibur.feature.testcase.dto.response.TestcaseDataDto;
 import com.sk.skala.axcalibur.feature.testcase.dto.response.TestcaseGenerationResponse;
 import com.sk.skala.axcalibur.feature.testcase.dto.response.TestcaseParamDto;
-import com.sk.skala.axcalibur.feature.testcase.entity.MappingEntity;
-import com.sk.skala.axcalibur.feature.testcase.entity.ParameterEntity;
-import com.sk.skala.axcalibur.feature.testcase.entity.ScenarioEntity;
 import com.sk.skala.axcalibur.feature.testcase.entity.TestCaseDataEntity;
 import com.sk.skala.axcalibur.feature.testcase.entity.TestCaseEntity;
 import com.sk.skala.axcalibur.feature.testcase.repository.MappingRepository;
@@ -28,6 +25,9 @@ import com.sk.skala.axcalibur.feature.testcase.repository.ParameterRepository;
 import com.sk.skala.axcalibur.feature.testcase.repository.TestCaseDataRepository;
 import com.sk.skala.axcalibur.feature.testcase.repository.TestCaseRepository;
 import com.sk.skala.axcalibur.global.code.ErrorCode;
+import com.sk.skala.axcalibur.global.entity.MappingEntity;
+import com.sk.skala.axcalibur.global.entity.ParameterEntity;
+import com.sk.skala.axcalibur.global.entity.ScenarioEntity;
 import com.sk.skala.axcalibur.global.exception.BusinessExceptionHandler;
 
 import feign.FeignException;
