+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/feature/testcase/service/TcPayloadService.java
@@ -3,7 +3,7 @@ package com.sk.skala.axcalibur.feature.testcase.service;
 import java.util.List;
 
 import com.sk.skala.axcalibur.feature.testcase.dto.request.TcRequestPayload;
-import com.sk.skala.axcalibur.feature.testcase.entity.ScenarioEntity;
+import com.sk.skala.axcalibur.global.entity.ScenarioEntity;
 
 /**
  * TC 생성 이전 단계의 서비스
