+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/global/entity/AvalonCookieEntity.java
@@ -1,4 +1,4 @@
-package com.sk.skala.axcalibur.feature.testcase.entity;
+package com.sk.skala.axcalibur.global.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
