+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/global/entity/CategoryEntity.java
@@ -1,6 +1,4 @@
-package com.sk.skala.axcalibur.feature.testcase.entity;
-
-import com.sk.skala.axcalibur.global.entity.BaseTimeEntity;
+package com.sk.skala.axcalibur.global.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -15,21 +13,21 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * 파라미터의 항목 유형 저장 테이블
- * 데이터베이스의 'context' 테이블과 매핑
+ * 데이터 파라미터의 항목 저장 테이블
+ * 데이터베이스의 'category' 테이블과 매핑
  */
 @Getter
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 @Entity
-@Table(name = "context")
-public class ContextEntity extends BaseTimeEntity {
+@Table(name = "category")
+public class CategoryEntity extends BaseTimeEntity{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "`key`")
-    private Integer id;         // 항목 유형 키(PK, AUTO_INCREMENT)
+    private Integer id;         // 파라미터 항목 키(PK, AUTO_INCREMENT)
 
     @Column(name = "name", nullable = false, unique = true)
-    private String name;        // 항목 유형 이름(body, header, query, path)
+    private String name;        // 파라미터 항목 이름(path/query, request, response)
 }
