+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/global/entity/ParameterEntity.java
@@ -1,4 +1,4 @@
-package com.sk.skala.axcalibur.feature.testcase.entity;
+package com.sk.skala.axcalibur.global.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
