+++ backend/scenario/src/main/java/com/sk/skala/axcalibur/global/entity/ScenarioEntity.java
@@ -1,6 +1,4 @@
-package com.sk.skala.axcalibur.feature.testcase.entity;
-
-import com.sk.skala.axcalibur.global.entity.BaseTimeEntity;
+package com.sk.skala.axcalibur.global.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
